bombbs
added solinium bomb and changed the rendering because it was fucked up
