+++ src/main/java/com/hbm/render/item/ItemRenderLibrary.java
@@ -363,29 +363,6 @@ public class ItemRenderLibrary {
 		        ResourceManager.bomb_boy.renderAll();
 			}});
 
-		renderers.put(Item.getItemFromBlock(ModBlocks.nuke_prototype), new ItemRenderBase() {
-			public void renderInventory() {
-				GL11.glScaled(2.25, 2.25, 2.25);
-			}
-			public void renderCommon() {
-				GL11.glRotated(90, 0, 1, 0);
-		        bindTexture(ResourceManager.bomb_prototype_tex);
-		        ResourceManager.bomb_prototype.renderAll();
-			}});
-
-		renderers.put(Item.getItemFromBlock(ModBlocks.nuke_solinium), new ItemRenderBase() {
-			public void renderInventory() {
-				GL11.glScaled(4, 4, 4);
-			}
-			public void renderCommon() {
-				GL11.glTranslated(0.5, 0, 0);
-				GL11.glRotated(90, 0, 1, 0);
-	            GL11.glDisable(GL11.GL_CULL_FACE);
-		        bindTexture(ResourceManager.bomb_solinium_tex);
-		        ResourceManager.bomb_solinium.renderAll();
-	            GL11.glEnable(GL11.GL_CULL_FACE);
-			}});
-
 		renderers.put(Item.getItemFromBlock(ModBlocks.nuke_n2), new ItemRenderBase() {
 			public void renderInventory() {
 				GL11.glTranslated(0, -4, 0);
