+++ src/main/java/com/hbm/render/tileentity/RenderNukeSolinium.java
@@ -1,5 +1,9 @@
 package com.hbm.render.tileentity;
 
+import com.hbm.blocks.ModBlocks;
+import com.hbm.render.item.ItemRenderBase;
+import net.minecraft.item.Item;
+import net.minecraftforge.client.IItemRenderer;
 import org.lwjgl.opengl.GL11;
 
 import com.hbm.main.ResourceManager;
@@ -7,8 +11,8 @@ import com.hbm.main.ResourceManager;
 import net.minecraft.client.renderer.tileentity.TileEntitySpecialRenderer;
 import net.minecraft.tileentity.TileEntity;
 
-public class RenderNukeSolinium extends TileEntitySpecialRenderer {
-	
+public class RenderNukeSolinium extends TileEntitySpecialRenderer implements IItemRendererProvider {
+
 	public RenderNukeSolinium() { }
 
     @Override
@@ -18,9 +22,9 @@ public class RenderNukeSolinium extends TileEntitySpecialRenderer {
         GL11.glTranslated(x + 0.5D, y, z + 0.5D);
         GL11.glEnable(GL11.GL_LIGHTING);
         GL11.glDisable(GL11.GL_CULL_FACE);
-        
+
         GL11.glRotatef(90, 0F, 1F, 0F);
-        
+
 		switch(tileEntity.getBlockMetadata())
 		{
 		case 2:
@@ -32,12 +36,39 @@ public class RenderNukeSolinium extends TileEntitySpecialRenderer {
 		case 5:
 			GL11.glRotatef(0, 0F, 1F, 0F); break;
 		}
-
+		GL11.glShadeModel(GL11.GL_SMOOTH);
         bindTexture(ResourceManager.bomb_solinium_tex);
         ResourceManager.bomb_solinium.renderAll();
-        
+		GL11.glShadeModel(GL11.GL_FLAT);
+
         GL11.glEnable(GL11.GL_CULL_FACE);
 
         GL11.glPopMatrix();
     }
+
+	@Override
+	public Item getItemForRenderer() {
+		return Item.getItemFromBlock(ModBlocks.nuke_solinium);
+	}
+
+	@Override
+	public IItemRenderer getRenderer() {
+		return new ItemRenderBase() {
+			public void renderInventory() {
+				GL11.glTranslated(0, -0.125, 0);
+				GL11.glScaled(5, 5, 5);
+			}
+
+			public void renderCommon() {
+				GL11.glRotated(90, 0, 1, 0);
+				GL11.glTranslated(0, -0.125, 0);
+				GL11.glShadeModel(GL11.GL_SMOOTH);
+				GL11.glDisable(GL11.GL_CULL_FACE);
+				bindTexture(ResourceManager.bomb_solinium_tex);
+				ResourceManager.bomb_solinium.renderAll();
+				GL11.glEnable(GL11.GL_CULL_FACE);
+				GL11.glShadeModel(GL11.GL_FLAT);
+			}
+		};
+	}
 }
